+++ site/components/ComponentList.js
@@ -201,6 +201,20 @@ class ComponentList {
             code: require("./samples/toast/ToastSample.txt")
         });
 
+        components.complex.push({
+            header: "ReCaptcha",
+            desc: "ReCaptcha",
+            sample: require("./samples/recaptcha/ReCaptchaSample"),
+            code: require("./samples/recaptcha/ReCaptchaSample.txt")
+        });
+
+        components.complex.push({
+            header: "GoogleMap",
+            desc: "is a wrapper component for GoogleMaps API",
+            sample: require("./samples/googlemap/GoogleMapSample"),
+            code: require("./samples/googlemap/GoogleMapSample.txt")
+        });
+
         components.charts.push({
             header: "AreaChart",
             desc: " is a layout chart",