+++ site/components/samples/charts/RadarChartSample.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { ShallowComponent } from "robe-react-commons";
-import { RadarChart } from "robe-react-ui/lib/chart";
+import RadarChart from "robe-react-ui/lib/chart/RadarChart";
 
 let data = [
     { name: "Page A", uv: 4000, pv: 2400, amt: 2400, fill: "#8884d8" },