+++ site/components/samples/charts/ScatterChartSample.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { ShallowComponent } from "robe-react-commons";
-import { ScatterChart } from "robe-react-ui/lib/chart";
+import ScatterChart from "robe-react-ui/lib/chart/ScatterChart";
 
 const data01 = [{ x: 100, y: 200, z: 200 }, { x: 120, y: 100, z: 260 },
                   { x: 170, y: 300, z: 400 }, { x: 140, y: 250, z: 280 },