+++ src/checktree/CheckTree.jsx
@@ -39,27 +39,29 @@ export default class CheckTree extends ShallowComponent {
         /**
          * Checked items array.
          */
-        value: React.PropTypes.array
+        value: React.PropTypes.array,
+        /**
+         * Valuen type
+         */
+        valueType: React.PropTypes.oneOf(["number", "string"])
     };
 
     static defaultProps = {
         textField: "text",
         valueField: "code",
         childrenField: "children",
-        value: []
+        value: [],
+        valueType: "string"
     };
 
     __value = [];
 
-    constructor(props: Object) {
+    constructor(props:Object) {
         super(props);
-        this.state = {
-            value: this.props.value
-        };
+        this.componentWillReceiveProps(props);
     }
 
-
-    render(): Object {
+    render():Object {
         return (
             <div>
                 <Tree
@@ -70,7 +72,7 @@ export default class CheckTree extends ShallowComponent {
                     valueField={this.props.valueField}
                     onChange={this.__onChange}
                     itemRenderer={CheckTree.itemRenderer}
-                    />
+                />
             </div>
         );
     }
@@ -82,7 +84,7 @@ export default class CheckTree extends ShallowComponent {
      * @returns {Object} a CheckInput
      * @memberOf CheckTree
      */
-    static itemRenderer(props: Object): Object {
+    static itemRenderer(props:Object):Object {
         let checked;
         let value = props.item[props.valueField];
         if (Arrays.indexOf(props.value, value) !== -1) {
@@ -97,10 +99,10 @@ export default class CheckTree extends ShallowComponent {
                 name={`${props.item[props.valueField]}`}
                 formControl={false}
                 onChange={props.onChange}
-                />);
+            />);
     }
 
-    __onChange = (e: Object) => {
+    __onChange = (e:Object) => {
         let values = this.state.value;
         let value = e.target.parsedValue[0];
 
@@ -125,7 +127,8 @@ export default class CheckTree extends ShallowComponent {
             this.props.onChange(values);
         }
     };
-    static filterDuplicates(src: Array): Array {
+
+    static filterDuplicates(src:Array):Array {
         let sorted = src.slice().sort();
         let result = [];
         if (sorted[0] !== undefined) {
@@ -139,12 +142,12 @@ export default class CheckTree extends ShallowComponent {
         return result;
     }
 
-    __getChildrenValues(selectedValue: any): Array {
+    __getChildrenValues(selectedValue:any):Array {
         let selectedChildren = [];
-        this.__traverseItems(this.props.items, (item: any) => {
+        this.__traverseItems(this.props.items, (item:any) => {
             let value = item[this.props.valueField];
             if (value === selectedValue) {
-                this.__traverseItems(item[this.props.childrenField], (item2: any) => {
+                this.__traverseItems(item[this.props.childrenField], (item2:any) => {
                     let value2 = item2[this.props.valueField];
                     selectedChildren.push(value2);
                 });
@@ -158,11 +161,11 @@ export default class CheckTree extends ShallowComponent {
      * @returns {Array}
      * @memberOf CheckTree
      */
-    getUnselectedItems(): Array {
+    getUnselectedItems():Array {
         let unselected = [];
         let items = this.props.items;
         let values = this.__value;
-        this.__traverseItems(items, (item: any) => {
+        this.__traverseItems(items, (item:any) => {
             let value = item[this.props.valueField];
             if (Arrays.indexOf(values, value) === -1) {
                 unselected.push(value);
@@ -176,11 +179,11 @@ export default class CheckTree extends ShallowComponent {
      * @returns {Array}
      * @memberOf CheckTree
      */
-    getSelectedItems(): Array {
+    getSelectedItems():Array {
         return this.__value;
     }
 
-    __traverseItems(items: Array, callback: Function) {
+    __traverseItems(items:Array, callback:Function) {
         if (items === undefined) {
             return;
         }
@@ -194,7 +197,13 @@ export default class CheckTree extends ShallowComponent {
     }
 
 
-    shouldComponentUpdate(): boolean {
+    shouldComponentUpdate():boolean {
         return true;
+    }
+
+    componentWillReceiveProps(nextProps) {
+        this.state = {
+            value: nextProps.value
+        };
     }
 }