+++ src/faicon/index.js
@@ -1,1 +1,2 @@
-export { default, default as FaIcon } from "./FaIcon";
+export { default, default as Icon } from "./Icon";
+export { default as FaIcon } from "./FaIcon";