+++ src/inputs/DateInput.jsx
@@ -1,5 +1,4 @@
 import React from "react";
-import Dom from "react-dom";
 import ShallowComponent from "robe-react-commons/lib/components/ShallowComponent";
 import momentjs from "moment";
 import is from "is-js";
@@ -106,9 +105,7 @@ export default class DateInput extends ShallowComponent {
         DateInput.idCounter++;
         this.state = {
             open: false,
-            value: this.props.value,
-            selectionStart: undefined,
-            selectionEnd: undefined
+            value: this.props.value
         };
         if (momentjs(this.props.value).isValid() && this.props.value !== "" && this.props.value !== undefined) {
             this.isPartial = false;
@@ -123,23 +120,26 @@ export default class DateInput extends ShallowComponent {
      * @returns
      */
     render():Object {
-        let parsedValue;
+        let parsedValue = "";
         let value = this.state.value;
-
         if (this.isPartial) {
             parsedValue = "Invalid date";
         } else if (is.number(value)) {
             parsedValue = momentjs(value).format(this.props.format);
         } else if (this.__checkPartialRegex(value)) {
             parsedValue = momentjs(value, this.props.format).format(this.props.format);
-            value = momentjs(value, this.props.format);
+            if (value)
+                value = momentjs(value, this.props.format);
         }
 
         let overlayValue;
         if (parsedValue === "Invalid date" || this.isPartial) {
             parsedValue = value;
         } else {
-            overlayValue = value === "" ? undefined : value;
+            overlayValue = value === "" ? undefined : parsedValue;
+        }
+        if (!is.number(overlayValue)) {
+            overlayValue = momentjs(overlayValue, this.props.format, true).toDate().getTime();
         }
         let {format, locale, minDate, maxDate, ...newProps} = this.props;
         return (
@@ -175,6 +175,10 @@ export default class DateInput extends ShallowComponent {
     }
 
     __onKeyDown(e) {
+        if (!(e.key === "ArrowUp" || e.key === "ArrowDown")) {
+            return;
+        }
+        e.preventDefault();
 
         let value = this.state.value;
         if (!value)
@@ -210,42 +214,33 @@ export default class DateInput extends ShallowComponent {
         }
 
         if (e.key === "ArrowUp") {
-            e.preventDefault();
-            valueParts[partIndex] = (parseInt(valueParts[partIndex]) + 1).toString();
+            let tempValue = parseInt(valueParts[partIndex]) + 1;
+            valueParts[partIndex] = tempValue < 10 ? '0' + tempValue : tempValue.toString();
         }
         else if (e.key === "ArrowDown") {
-            e.preventDefault();
             let tempValue = parseInt(valueParts[partIndex]) - 1;
-            valueParts[partIndex] = (tempValue <= 0 ? "01" : tempValue).toString();
-        } else {
-            this.setState({selectionStart: undefined, selectionEnd: undefined});
-            return;
+            valueParts[partIndex] = tempValue < 10 ? '0' + tempValue : tempValue.toString();
         }
         value = valueParts.join(this.separator);
         let valid = momentjs(value, this.props.format).format(this.props.format);
 
         if (!this.__checkPartialRegex(value) || valid === "Invalid date") {
-            this.setState({selectionStart: undefined, selectionEnd: undefined});
             return;
         }
         this.setState({
             value: value,
-            selectionStart: selectionStart,
-            selectionEnd: selectionEnd,
             open: false
         });
-
         e.target.parsedValue = momentjs(value, this.props.format).toDate().getTime();
         e.target.value = value;
         if (this.props.onChange)
             this.props.onChange(e);
+
+        e.target.selectionStart = selectionStart;
+        e.target.selectionEnd = selectionEnd;
     }
 
     __onKeyUp(e) {
-        if (this.state.selectionStart != undefined && this.state.selectionEnd != undefined) {
-            e.target.selectionStart = this.state.selectionStart;
-            e.target.selectionEnd = this.state.selectionEnd;
-        }
     }
 
     /**
@@ -276,11 +271,11 @@ export default class DateInput extends ShallowComponent {
         e.target.name = this.props.name;
 
         if (!this.validChars.test(value) || !this.__checkPartialRegex(value)) {
+            // Do not take input if maxlength exeeded or invalid char entered.
             this.setState({
                 color: "#a94442"
             });
-            return true;
-        } else if (this.props.onChange) {
+        } else if (this.props.onChange && value.length <= this.props.format.length) {
             this.setState({
                 color: undefined
             });
@@ -292,13 +287,14 @@ export default class DateInput extends ShallowComponent {
                     open: false,
                     value: e.target.value
                 });
+                if (e.target.value === "")
+                    this.props.onChange(e);
                 return;
             }
             result = this.props.onChange(e);
         }
         this.setState({
-            open: false,
-            value: e.target.value
+            open: false
         });
 
         if (!result) {