+++ src/recaptcha/index.js
@@ -1,0 +1,1 @@
+export { default, default as ReCaptcha } from "./ReCaptcha";